Autodetect CSV delimiter
This uses `csv-sniffer` to autodetect the CSV delimiter. It does so by adding a
`PassThrough()` stream that reads the first chunk of data, infers the
delimiter, and configures the existing CSV parser.

I hoped to be able to do this without creating a dependency between the CSV
delimiter extraction and the CSV parser, but was unable to because the pipeline
needs to be configured before reading, otherwise the CSV parser won't get the
first chunk.

I also tried adding a test to ensure that if an explicit delimiter is passed,
it'll be used, but failed because for some reason I couldn't catch an error
thrown by `csv-parser`. Didn't matter that the code was inside a try/catch
block, the error was still raised.